New version of GenerativeArt.py
Implemented the managment of the layer's elements. Now each layer can have different ammounts of elements.
Reformatting the paths.txt and updated the readme.txt. 
Added the generation of a log file which will containt all the informations abaout exectuing time and number of iteration done for generating all the NFTs.